fix: alerts and filters
