+++ frontend/src/queries/systemAlerts/systemAlertHistory.ts
@@ -17,6 +17,7 @@ export const useSystemAlertHistory = defineQuery(() => {
   const sortDescending = ref(true)
   const severityFilters = ref<string[]>([])
   const alertnameFilters = ref<string[]>([])
+  const statusFilters = ref<string[]>([])
 
   const { state, asyncStatus, ...rest } = useQuery({
     key: () => [
@@ -28,6 +29,7 @@ export const useSystemAlertHistory = defineQuery(() => {
       sortDescending.value,
       severityFilters.value.join(','),
       alertnameFilters.value.join(','),
+      statusFilters.value.join(','),
     ],
     enabled: () => !!loginStore.jwtToken && !!route.params.systemId,
     query: () =>
@@ -39,15 +41,17 @@ export const useSystemAlertHistory = defineQuery(() => {
         sortDescending.value,
         severityFilters.value.length > 0 ? severityFilters.value : undefined,
         alertnameFilters.value.length > 0 ? alertnameFilters.value : undefined,
+        statusFilters.value.length > 0 ? statusFilters.value : undefined,
       ),
   })
 
   const areDefaultFiltersApplied = () =>
-    !severityFilters.value.length && !alertnameFilters.value.length
+    !severityFilters.value.length && !alertnameFilters.value.length && !statusFilters.value.length
 
   const resetFilters = () => {
     severityFilters.value = []
     alertnameFilters.value = []
+    statusFilters.value = []
     pageNum.value = 1
   }
 
@@ -61,6 +65,7 @@ export const useSystemAlertHistory = defineQuery(() => {
     sortDescending,
     severityFilters,
     alertnameFilters,
+    statusFilters,
     areDefaultFiltersApplied,
     resetFilters,
   }
